Начало доработок для сборки mingw

+++ crate_windows.go
@@ -1,3 +1,6 @@
+//go:build windows
+// +build windows
+
 package ltrgo
 
 /*
@@ -37,9 +40,9 @@ func GetCrateSerials() ([]string, error) {
 		return nil, ErrOpenCrate
 	}
 	foundCrates := 0
-	foundCratesPtr := (*C.uint)(unsafe.Pointer(&foundCrates))
+	foundCratesPtr := (*C.ulong)(unsafe.Pointer(&foundCrates))
 	cratesReturned := 0
-	cratesReturnedPtr := (*C.uint)(unsafe.Pointer(&cratesReturned))
+	cratesReturnedPtr := (*C.ulong)(unsafe.Pointer(&cratesReturned))
 	var infoList [C.LTR_CRATES_MAX]C.TLTR_CRATE_INFO
 	infoListPtr := (*C.TLTR_CRATE_INFO)(unsafe.Pointer(&infoList))
 	var serials [C.LTR_CRATES_MAX][C.LTR_CRATE_SERIAL_SIZE]C.char
